Added Storage class. Organized methods and variables that deal with loading and saving into Storage class.
